+++ tests/all_tests.rs
@@ -2,7 +2,6 @@
 #![cfg_attr(feature = "adt_const_marker", feature(adt_const_params))]
 #![cfg_attr(feature = "adt_const_marker", feature(unsized_const_params))]
 #![cfg_attr(feature = "adt_const_marker", allow(incomplete_features))]
-#![cfg_attr(feature = "nightly_mut_refs", feature(const_mut_refs))]
 
 mod misc_tests {
     mod test_utils;